+++ examples/base_1cc/variables.tf
@@ -92,12 +92,17 @@ variable "azure_vault_url" {
 variable "ccvm_instance_type" {
   type        = string
   description = "Cloud Connector Image size"
-  default     = "Standard_D2s_v3"
+  default     = "Standard_D2ds_v5"
   validation {
     condition = (
       var.ccvm_instance_type == "Standard_D2s_v3" ||
       var.ccvm_instance_type == "Standard_DS2_v2" ||
-      var.ccvm_instance_type == "Standard_DS3_v2"
+      var.ccvm_instance_type == "Standard_D2ds_v4" ||
+      var.ccvm_instance_type == "Standard_D2ds_v5" ||
+      var.ccvm_instance_type == "Standard_DS3_v2" ||
+      var.ccvm_instance_type == "Standard_D8s_v3" ||
+      var.ccvm_instance_type == "Standard_D16s_v3" ||
+      var.ccvm_instance_type == "Standard_DS5_v2"
     )
     error_message = "Input ccvm_instance_type must be set to an approved vm size."
   }