+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## TBD UNRELEASED
+
+ENHANCEMENTS:
+* add: Standard_D2ds_v4 and Standard_D2ds_v5 size support
+* change: Default recommended small CC instance to Standard_D2ds_v5
+
 ## v0.7.0 (March 10, 2025)
 
 FEATURES: